+++ src/app/app-routing.module.ts
@@ -3,7 +3,6 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 //Components
 import { NotFoundComponent } from './components/not-found/not-found.component';
-import { FavListComponent } from './components/fav-list/fav-list.component';
 import { ServerErrorComponent } from './components/server-error/server-error.component';
 import { HomeComponent } from './components/home/home.component';
 
@@ -25,11 +24,6 @@ const routes: Routes = [
       import('./modules/food-module/food-module.module').then(
         (module) => module.FoodModuleModule
       ),
-  },
-  //Favorite recipies list route
-  {
-    path: 'fav-list',
-    component: FavListComponent,
   },
   //Server error route
   { path: 'server-error', component: ServerErrorComponent },
