+++ src/app/app.module.ts
@@ -1,6 +1,9 @@
 //Modules
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { ServiceWorkerModule } from '@angular/service-worker';
+import { StoreModule } from '@ngrx/store';
+import { environment } from '../environments/environment';
 import { AppRoutingModule } from './app-routing.module';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
@@ -19,26 +22,23 @@ import { DataViewModule } from 'primeng/dataview';
 import { HttpClientModule } from '@angular/common/http';
 import { PasswordModule } from 'primeng/password';
 import { DropdownModule } from 'primeng/dropdown';
+//Reducers
+import { recipiesReducer } from './store/food-recipies/food-recipies.reducers';
+import { favouritesReducer } from './store/favourite-list/fav-list.reducers';
+//Interceptors
+import { interceptorProviders } from './services/interceptors/interceptor-providers';
 //Components
 import { AppComponent } from './app.component';
 import { NavBarComponent } from './components/nav-bar/nav-bar.component';
 import { NotFoundComponent } from './components/not-found/not-found.component';
-import { FavListComponent } from './components/fav-list/fav-list.component';
-import { StoreModule } from '@ngrx/store';
-import { recipiesReducer } from './store/food-recipies/food-recipies.reducers';
-import { favouritesReducer } from './store/favourite-list/fav-list.reducers';
 import { ServerErrorComponent } from './components/server-error/server-error.component';
-import { ServiceWorkerModule } from '@angular/service-worker';
-import { environment } from '../environments/environment';
 import { HomeComponent } from './components/home/home.component';
-import { interceptorProviders } from './services/interceptor/interceptor-providers';
 
 @NgModule({
   declarations: [
     AppComponent,
     NavBarComponent,
     NotFoundComponent,
-    FavListComponent,
     ServerErrorComponent,
     HomeComponent,
   ],
