+++ src/app/modules/auth-module/components/login/login.component.scss
@@ -1,5 +1,5 @@
 .background-img {
-  background-image: url("../../../../assets/images/login-background.jpg");
+  background-image: url("../../../../../assets/images/login-background.jpg");
   height: 92.2vh;
   background-size: cover;
   background-position: center;
