+++ src/app/modules/auth-module/components/register/register.component.scss
@@ -1,5 +1,5 @@
 .background-img {
-  background-image: url("../../../../assets/images/register-background.jpg");
+  background-image: url("../../../../../assets/images/register-background.jpg");
   background-size: cover;
   background-position: center;
   background-repeat: no-repeat;
