+++ src/app/modules/auth-module/components/register/register.component.ts
