+++ src/app/modules/food-module/components/categories/categories.component.html
@@ -14,7 +14,6 @@
         <p class="visually-hidden">Loading...</p>
       </div>
     </div>
-
     <!-- Taps -->
     <!-- Pizza tab -->
     <p-tabPanel header="Pizza" [selected]="true">
