+++ src/app/modules/food-module/components/food-card/foodcard.component.html
@@ -1,24 +1,23 @@
-<p-card styleClass="p-card-shadow" *ngIf="foodData" >
+<p-card styleClass="p-card-shadow" *ngIf="foodData">
   <!-- Recipe image -->
   <ng-template pTemplate="header">
     <figure>
-      <img alt="Card" [src]="foodData.image_url" class="rounded img-fluid"/>
+      <img alt="Card" [src]="foodData.image_url" class="rounded img-fluid" />
     </figure>
   </ng-template>
-
   <!-- Recipe title -->
   <div class="card-data">
     <h5>{{ foodData.title }}</h5>
     <h6 class="my-4">{{ foodData.publisher }}</h6>
   </div>
-
-  
   <!-- Recipe details -->
   <ng-template pTemplate="footer">
     <div class="d-flex justify-content-between">
       <p class="my-auto">Rate: {{ foodData.social_rank.toFixed() }}%</p>
       <!-- Save to fav button -->
-      <button type="button" class="favButton "><i class="pi pi-bookmark"></i></button>
+      <button type="button" class="favButton">
+        <i class="pi pi-bookmark"></i>
+      </button>
     </div>
   </ng-template>
 </p-card>
