+++ src/app/modules/food-module/components/food-card/foodcard.component.scss
@@ -1,22 +1,21 @@
-:host ::ng-deep .p-card{
-    box-shadow: 5px 10px 18px #888888;
-    height: 34.375rem;
+:host ::ng-deep .p-card {
+  box-shadow: 5px 10px 18px #888888;
+  height: 34.375rem;
 }
-.favButton{
-    background-color: #fff;
-    border: none;
-    i{
-        color: #ea4645;
-        font-size: 1.5rem;
-    }
+.favButton {
+  background-color: #fff;
+  border: none;
+  i {
+    color: #ea4645;
+    font-size: 1.5rem;
+  }
 }
 img {
   height: 18.75rem;
 }
-:host ::ng-deep .p-card .p-card-body{
-    padding-top: 0;
+:host ::ng-deep .p-card .p-card-body {
+  padding-top: 0;
 }
-.card-data{
-    height: 100px;
+.card-data {
+  height: 100px;
 }
-
