+++ src/app/modules/food-module/components/food-card/foodcard.component.ts
@@ -4,13 +4,11 @@ import { Recipe } from 'src/app/interfaces/food-category';
 @Component({
   selector: 'app-foodcard',
   templateUrl: './foodcard.component.html',
-  styleUrls: ['./foodcard.component.scss']
+  styleUrls: ['./foodcard.component.scss'],
 })
 export class FoodcardComponent implements OnInit {
-  
-  @Input() foodData:Recipe | null =null;
-  constructor() { }
+  @Input() foodData: Recipe | null = null;
+  constructor() {}
 
   ngOnInit(): void {}
-
 }
