+++ src/app/modules/food-module/food-module-routing.module.ts
@@ -4,6 +4,7 @@ import { RouterModule, Routes } from '@angular/router';
 //Components
 import { FoodrecipeComponent } from './components/recipe-details/foodrecipe.component';
 import { CategoriesComponent } from './components/categories/categories.component';
+import { FavListComponent } from './components/fav-list/fav-list.component';
 
 const routes: Routes = [
   //Default route
@@ -12,6 +13,11 @@ const routes: Routes = [
   { path: 'categories', component: CategoriesComponent },
   //Recipe details route
   { path: 'recipe/:id', component: FoodrecipeComponent },
+  //Favorite recipies list route
+  {
+    path: 'fav-list',
+    component: FavListComponent,
+  },
 ];
 
 @NgModule({
