--- src/app/services/interceptor/language/language-interceptor.service.ts
@@ -1,18 +1,0 @@
-import { HttpEvent, HttpHandler, HttpInterceptor, HttpRequest } from '@angular/common/http';
-import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
-@Injectable({
-  providedIn: 'root'
-})
-export class LanguageInterceptorService implements HttpInterceptor {
-
-  constructor() { }
-  intercept(request: HttpRequest<any>,next: HttpHandler): Observable<HttpEvent<any>> {
-    request = request.clone({
-      
-      setHeaders:{'Accept-Language':'en'},
-      setParams:{'lang':'en'}
-    });
-    return next.handle(request);
-  }
-}
