+++ src/app/store/food-recipies/food-recipies.reducers.ts
@@ -13,10 +13,9 @@ export const recipiesReducer = createReducer(
   initialState,
   //Fill recipes action
   on(fillRecipies, (state, action) => {
-    console.log(action);
     return {
       ...state,
-      recipes: [...action.payload.recipes],
+      recipes: [...state.recipes, ...action.payload.recipes],
       category: action.payload.category,
     };
   }),
